requirements.txt
asgiref==3.8.1
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-template-partials==23.4
djhtml==3.0.6
gunicorn==21.2.0
packaging==24.0
pillow==10.3.0
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.6.0
